[Gold IV] Title: 최소 스패닝 트리, Time: 560 ms, Memory: 48916 KB -BaekjoonHub

+++ 백준/Gold/1197. 최소 스패닝 트리/최소 스패닝 트리.java
@@ -32,7 +32,7 @@ public class Main {
 		Arrays.sort(edges);
 		
 		int totalEdges = 0;
-		int minWeight = 0;
+		long minWeight = 0;
 		for(int i = 0; i < E; i++) {
 			Edge edge = edges[i];
 			int a = edge.a;
